# backend/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    name: DevHelp Team
    url: https://github.com/bloiss/DevHelp
  description: API REST du forum DevHelp — entraide Dev & IA.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DevHelp API
  version: "1.0"
paths:
...
  /api/v1/notifications:
    get:
      description: Retourne toutes les notifications de l'utilisateur connecté.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lister toutes les notifications
      tags:
      - notifications
  /api/v1/notifications/{id}:
    delete:
      description: Supprime définitivement une notification de l'utilisateur connecté.
      parameters:
      - description: UUID notification
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Supprimer une notification
      tags:
      - notifications
  /api/v1/notifications/{id}/archive:
    patch:
      consumes:
      - application/json
      description: Déplace une notification dans les archives ou la restaure.
      parameters:
      - description: UUID notification
        in: path
        name: id
        required: true
        type: string
      - description: 'archived: true/false'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Archiver ou désarchiver une notification
      tags:
      - notifications
  /api/v1/notifications/{id}/read:
    patch:
      description: Marque une notification spécifique comme lue.
      parameters:
      - description: UUID notification
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Marquer une notification comme lue
      tags:
      - notifications
  /api/v1/notifications/{id}/star:
    patch:
      consumes:
      - application/json
      description: Marque ou démarque une notification avec une étoile pour la mettre
        en favori.
      parameters:
      - description: UUID notification
        in: path
        name: id
        required: true
        type: string
      - description: 'starred: true/false'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Étoiler ou désétoiler une notification
      tags:
      - notifications
  /api/v1/notifications/{id}/unread:
    patch:
      description: Marque une notification spécifique comme non lue.
      parameters:
      - description: UUID notification
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Marquer une notification comme non lue
      tags:
      - notifications
  /api/v1/notifications/inbox:
    get:
      description: Retourne les notifications non archivées de l'utilisateur connecté
        (inbox).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Boîte de réception des notifications
      tags:
      - notifications
  /api/v1/notifications/prefs:
    get:
      description: Retourne les préférences de notifications push de l'utilisateur
        connecté.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Récupérer les préférences de notifications
      tags:
      - notifications
    patch:
      consumes:
      - application/json
      description: Met à jour les préférences de notifications push (push activé,
        commentaires, likes, messages).
      parameters:
      - description: Préférences (push_enabled, notify_on_comment, notify_on_like,
          notify_on_message)
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mettre à jour les préférences de notifications
      tags:
      - notifications
  /api/v1/notifications/read-all:
    patch:
      description: Marque toutes les notifications de l'utilisateur connecté comme
        lues.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Marquer toutes les notifications comme lues
      tags:
      - notifications
  /api/v1/notifications/unread-count:
    get:
      description: Retourne le nombre de notifications non lues de l'utilisateur connecté.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Nombre de notifications non lues
      tags:
      - notifications
...
  /api/v1/push/subscriptions:
    delete:
      consumes:
      - application/json
      description: Supprime un abonnement Web Push enregistré pour ne plus recevoir
        de notifications sur cet endpoint.
      parameters:
      - description: Endpoint de la subscription à supprimer
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Supprimer un abonnement push
      tags:
      - notifications
    post:
      consumes:
      - application/json
      description: Enregistre un abonnement Web Push pour recevoir des notifications
        en temps réel.
      parameters:
      - description: Subscription (endpoint, p256dh, auth)
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Enregistrer un abonnement push
      tags:
      - notifications
  /api/v1/push/vapid-key:
    get:
      description: Retourne la clé publique VAPID nécessaire pour s'abonner aux notifications
        push.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Récupérer la clé publique VAPID
      tags:
      - notifications
...
securityDefinitions:
  BearerAuth:
    description: Entrez "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"